+++ src/main/java/chornarin/com/kh/Phone_Shop/Exception/ResourceNotFoundException.java
@@ -6,7 +6,7 @@ public class ResourceNotFoundException extends ApiException {
 
 
 
-    public ResourceNotFoundException(String Resourcename, Integer id) {
+    public ResourceNotFoundException(String Resourcename, Long id) {
         super(HttpStatus.NOT_FOUND, String.format("%s with id =  %d not found", Resourcename, id));
     }
