+++ src/main/java/chornarin/com/kh/Phone_Shop/Repository/ModelRepository.java
@@ -4,10 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import chornarin.com.kh.Phone_Shop.models.PhoneModel;
-import java.util.List;
 
 @Repository
-public interface ModelRepository extends JpaRepository<PhoneModel, Integer > {
+public interface ModelRepository extends JpaRepository<PhoneModel, Long > {
 
     PhoneModel findByBrandId(Integer brandId);
